Updated style management
