+++ src/empty-state/index.jsx
@@ -2,21 +2,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Radium from 'radium';
 import pure from 'recompose/pure';
-import styles from '../style/empty-states';
+import styles from './styles';
+import getStyles from './get-styles';
 import combineStyles from '../internal/combine-styles';
-
-function getStyle(image, overrideStyle) {
-  const style = combineStyles(styles.emptyState, { backgroundImage: `url(${image})` });
-
-  return combineStyles(style, overrideStyle);
-}
 
 /** Pretty placeholder for empty content */
 function EmptyState({
   headerText, bodyText, button, background, style, headingStyle, bodyStyle, ...custom
 }) {
   return (
-    <section style={getStyle(background, style)} {...custom}>
+    <section style={getStyles.root(background, style)} {...custom}>
       <h1 style={combineStyles(styles.heading, headingStyle)}>{headerText}</h1>
       <p style={combineStyles(styles.body, bodyStyle)}>{bodyText}</p>
       {button}
