+++ src/icon-menu/index.jsx
@@ -4,7 +4,7 @@ import Radium from 'radium';
 import pure from 'recompose/pure';
 import PopOver from '../pop-over';
 import Button from '../button';
-import styles from '../style/icon-menu';
+import styles from './styles';
 import combineStyles from '../internal/combine-styles';
 import getPopOverPosition from '../internal/get-pop-over-position';
