+++ src/message-input/index.jsx
@@ -3,27 +3,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Radium from 'radium';
 import shallowEqual from 'recompose/shallowEqual';
-import styles from '../style/message-inputs';
+import styles from './styles';
+import getStyles from './get-styles';
 import Button from '../button';
 import IconSend from '../icons/icon-send';
 import colors from '../settings/colors';
 import combineStyles from '../internal/combine-styles';
-
-function getButtonStyle(style, disabled) {
-  if (disabled) {
-    return combineStyles(style, styles.disabled);
-  }
-
-  return style;
-}
-
-function getInputStyle(rightButton, overrideStyle) {
-  if (rightButton) {
-    return combineStyles(combineStyles(styles.messageInput, styles.leftButton), overrideStyle);
-  }
-
-  return combineStyles(styles.messageInput, overrideStyle);
-}
 
 /** Message input with send button */
 class MessageInput extends Component {
@@ -108,13 +93,13 @@ class MessageInput extends Component {
       <section style={combineStyles(styles.input, style)}>
         {
           leftButton
-          ? <div style={getButtonStyle(styles.button, disabled)}>
+          ? <div style={getStyles.button(styles.button, disabled)}>
             {leftButton}
           </div>
           : null
         }
         <input
-          style={getInputStyle(leftButton, inputStyle)}
+          style={getStyles.input(leftButton, inputStyle)}
           placeholder={placeholder}
           onChange={onChange}
           value={value}
@@ -127,7 +112,7 @@ class MessageInput extends Component {
           {...custom}
         />
         <Button
-          style={getButtonStyle(styles.rightButton, disabled)}
+          style={getStyles.button(styles.rightButton, disabled)}
           iconButton
           onClick={sendMessage}
         >
